test(sqlite): add in-memory test harness seeded via migrations
